Re-add ContentAddressableStorage.
This interface will not be used by bb-storage directly, but by bb-worker
and bb-browser instead. Placing it in this repo makes sense, because it
provides symmetry with pkg/ac/action_cache.go.

While there, import the UncachedActionResult message. The name is
different compared to what's in the old Buildbarn tree. This is
intentional, as the old one was a misnomer.