Allow bb-storage to be used as a simple build cache.
To appease Bazel in setups like these, we'll need to provide an
implementation of GetCapabilities(). This call is normally implemented
by Buildbarn's scheduler process. For simple cached, we should let
bb-storage return a simple response instead.

Add a NonExecutingBuildQueue that acts as a simple stub that only
implements GetCapabilities. Make these part of the schedulers map on
startup. Also extend the ActionCacheServer, so that we can use the set
of supported instance names to explictly whitelist write access.