Replace FileInfo.Mode() by FileInfo.Type().
We don't need to have access to the actual UNIX permission bits in any
part of Buildbarn. We only need to know whether the file is executable.
Replace the mode by a simple enumeration.

Because FileInfo is now no longer a subset of os.FileInfo, remove the
interface. Merge it with SimpleFileInfo.