Support partial read requests from CAS
Bazel 0.26.0 seems to require support for partial reads with an offset
but no limit if a previous read of the same blob returned only a prefix
of the contents and then had some failure.

The implementation is generic and reads the blob from the start and just
throws away some data. This could be improved by changing the interfaces
so that each storage backend can handle the situation itself. Note that
the MerkleBlobAccess, that validates the content of the requested blobs,
requires reading from the beginning, so such optimization might not be
worth it depending on how often partial reads occur.

Fixes #15.