Introducing the buffer layer
The existing BlobAccess API uses io.ReadClosers to transfer object
contents. This has various disadvantages:

- It prevents us from supporting random access I/O (offset reads).
- There are various code paths that perform unnecessary conversions from
  byte slices to readers to byte slices, thereby copying the data.
- It makes it hard to provide high-level features such as retrying,
  error handling, metrics capturing, etc.

This change adds a new type called Buffer. Buffers can be created using
New{Kind}BufferFrom{Type}(). For example: NewCASBufferFromByteSlice().
The Buffer interface then provides a set of functions that allow the
buffer to be converted to the desired format. This means that if a
buffer is backed by a byte slice and ToByteSlice() is called on it, no
conversions take place whatsoever.