Add a utility library for pluggable eviction algorithms
There are some types in the codebase, such as
HardlinkingContentAddressableStorage, that handroll their own cache
replacement algorithm. These currently use random replacement. For
certain workloads, it would be preferred to use other algorithms, such
as LRU.

This change adds a simple interface for a set of strings (identifiers).
The set has a special function to report the object that should be
removed preferably.