Jsonnet: make the environment available through std.extVar()
It may sometimes be useful to be able to pass in configuration options
of Buildbarn binaries through environment variables. Examples include
running Buildbarn on top of Kubernetes, where you want to use the name
of a pod or node in the configuration. Kubernetes can be configured to
pass those in as environment variables. These can now be extracted using
std.extVar().