Make the filesystem API type safe
Instead of having an OpenFile() function that returns a single interface
for all file handles, create dedicated functions for readable, writable,
appendable and read-writable descriptors.

This not only prevents foot-shooting. It is also intended to make it
easier to make custom implementations of Directory without worrying
about POSIX semantics too much.

Files opened for reading now only support random access. This makes it
safe to access them in parallel. This allows us to implement PutFile()
in such a way that we don't need to seek.