Add a unified API for creating gRPC clients and servers
Introduce a common Protobuf message for configuring gRPC clients and
servers. Servers can listen on multipel TCP and UNIX socket addresses.
They may use TLS with optional client certificate authentication.

The goal is that we're going to use these functions uniformally
throughout the Buildbarn codebase, so that all of our components have
the same set of features when it comes to using gRPC.