Properly convert context errors (timeouts, cancelation) to gRPC errors
What is slightly annoying about gRPC is that errors returned by the
context package are not automatically converted to ones that have the
expected gRPC return code. For example, context.Canceled is not
converted to a CANCELED gRPC automatically.

This change adds a utility function that we can use throughout the
Buildbarn codebase to convert context errors to proper gRPC ones.