Add a BlobAccess decorator for coherence between the CAS and AC
When using Builds without the Bytes, it is expected that a
GetActionResult() only returns valid results if all of the blobs
referenced by it are actually contained within the CAS. Furthermore, it
must ensure that those objects don't disappear at a later point during
the build.

This decorator wraps any access to the AC in such a way that it computes
the transitive closure of referenced CAS blobs. It then calls
FindMissingBlobs() on those, so that their timestamp may be bumped.

This decorator needs to be enabled within a layer within the storage
stack that has access to all of the data (so not on a single storage
shard). This would preferably be the lowest layer, so that latency is
kept to a minimum.