Add LocalBlobAccess: a backend for storing data on the local system
The CircularBlobAccess storage backend was a nice experiment, but it has
a couple of serious flaws:

- There is a data corruption that may occur when reading blobs that are
  about to be overwritten by new content.
- It implements a FIFO-like eviction scheme, which is completely
  incompatible with Builds without the Bytes.

This change introduces LocalBlobAccess, a new storage backend that
attempts to fix this. The first problem is solved by storing data in a
series of files (blocks), instead of using a single file. Instead of
overwriting data, blocks are replaced by brand new copies.

This on its own would still cause storage to be FIFO. To solve that,
additional logic is added that causes data to be copied from older
blocks to newer blocks upon access. We don't want to do this too
frequently, because that would make accessing storage expensive. To
solve that, a threshold is used to determine how far an object has to be
within storage for it to be eligible to refreshes.

Right now this backend only supports in-memory storage. We should work
towards extending this to support disk-backed storage. This would allow
us to kick out the circular storage backend.