Add a mirrored storage backend
This class effectively implements RAID 1 for storage servers. It
provides a decorator for BlobAccess that replicates data lazily when
inconsistencies are detected.

- Put() calls go to both backends.
- Get() calls will access backends in a round-robin fashion. When
  observing NotFound errors, a fallback to the secondary backend is
  performed. If that were to return success, a synchronization takes
  place.
- FindMissing() is called against both backends. The intersection is
  returned and blobs are synchronized in both directions where
  necessary.

These semantics should be sufficient to support Builds without the
Bytes. It does require that both servers in the pair remain accessible,
but it is perfectly fine to replace a dead instance by an empty
replacement.