Don't announce UpdateEnabled == true for non-writable AC instance names
Buildbarn would erroneously announce that the AC is writable, even for
instance names for which this isn't explicitly enabled in the
configuration. This might cause Bazel to call PutActionResult(), even
though it is not permitted to do this.

Turn the default of UpdateEnabled to false. Add a decorator that
explicitly flips this to true for instance names for which this is
permitted in the configuration.