Let UncachedActionResult store ExecuteResponses
We currently let those store just the ActionResult. The downside of that
is that certain fields, most notable the Status result field is omitted.
This means that bb-browser would sometimes show an action as having
completed successfully, even though it did fail horrendously.

By storing an ExecuteResponse, we're able to store the Status result and
the server logs, so that bb-browser may display them and show the action
as actually failing, as it is.