Add an extension to the REv2 to store arbitrary execution metadata
bb_runner and bb_worker now have extensions that allow them to attach
file pool stats and POSIX rusage to ActionResult messages. This is
achieved by adding a repeated google.protobuf.Any field to
ExecutedActionMetadata.

We should attempt to send a change similar to this one upstream soon, so
that REv2 may support this out of the box.