Make it possible to set up keepalives on gRPC clients
Network stacks tend to be quite conservative when it comes to treating
connections as dead. It may often take far too long for them to notice
the remote side has disappeared.

gRPC supports keepalives, which may be used to detect this more quickly.
This feature seems to be turned off by default, which may make sense if
bandwidth is a concern. Let's go ahead and add a direct binding for the
API, so that we can enable it how we see fit.