Make completeness checking configurable as an ordinary BlobAccess
This feature was initially implemented not as an implementation of
BlobAccess. This meant that it made sense to have a boolean flag to
enable/disable this functionality.

Now that it is a BlobAccess, there is no advantage to having it as a
boolean flag. In fact, it's a disadvantage, because it means that other
bb_* binaries cannot have this functionality enabled.