Update the dependency on grpc-go
The latest version of grpc-go has integrated support for installing
chained interceptors, which is useful to have if we want to add support
for credential forwarding.

Upgrading this dependency causes one of the existing unit tests to fail,
due to status.Convert(nil).Proto() now returning nil instead of an empty
Status message. Patch up the unit test accordingly.