Allow specifying the gRPC keepalive enforcement policy
If a gRPC client uses a keepalive configuration with tight bounds, there
is a chance that it's too strict for a gRPC server to be willing to
respect it. This causes the client to exponentially back off until
bounds are obtained that work.

This change makes it possible to actually use tight keepalive bounds by
making the gRPC server's keepalive enforcement policy adjustable through
the configuration files.