Add a dedicated system for generic configuration options
There are certain configuration options that basically apply to all
Buildbarn binaries. One example is Jaeger. Right now only bb_storage
supports it, even though we want it to be configurable for all binaries.

Introduce a separate set of messages for setting up these options. Add a
pkg/global package that contains an initialization function for all of
these. From now on we'll be calling this from within main() in all
binaries.

This should make it easier to complete this PR:

https://github.com/buildbarn/bb-remote-execution/pull/32