Add support for grpc health check protocol
This is used with a probe like
https://github.com/grpc-ecosystem/grpc-health-probe or
https://www.envoyproxy.io/docs/envoy/latest/api-v2/api/v2/core/health_check.proto#envoy-api-msg-core-healthcheck-grpchealthcheck
to determine whether a service is both healthy and the correct
service. It is typically deployed as an load balancer health check, to
handle the scenario where:

  1. Our service is running on 1.2.3.4:8981
  2. Our service is rescheduled to run on 2.3.4.5:8981
  3. A different grpc server is started on 1.2.3.4:8981

Named service health checks allow the load balancer to notice that it
has the wrong backend, and not send it any rpcs.

In a future iteration, we should also redesign the API in this package
so that the various services can annnounce when they are ready to
serve traffic, for zero-downtime rollouts behind a load balancer. For
now, just serving a named health check at all is a step forwards.