WIP: Add a persistent DigestLocationMap
Adds a persistent digest-location map, works pretty much the same way as
the in-memory one. As the "cuckoo" hashing approach of the keys mean
that we end up with an integral key, we may simply use this to define an
offset in a file, that is

    offset = key * serialised object width

I've made this state directory configurable, and only be created if this
is set. I'm not particularly happy with this, and I'd like advice on a
better way to specify this.