WIP: Store LocalBlobAccess state for block device
Adds state storage for local blob access in the case it's using a
BlockDevice backend. At the moment this means that the BlockDevice
backend is always stateful, which may not necessarily be desirable. The
state is stored in a kind of weird format where we have several lines
containing the position of the block, what kind of block it is and the
offset it's at. This gets reloaded by a LocalBlobAccess on startup if
the file exists.