Don't store the empty blob explicitly
The REv2 spec has been extended to mention that clients never need to
upload the empty blob. They can assume that this blob is always present.

This change adds a decorator for BlobAccess that is automatically added
that intercepts Get(), Put() and FindMissing() calls for empty blobs.
This makes all Buildbarn components behave similarly to Bazel.

Fixes: #56