Add an option to disable checksum validation for LocalBlobAccess
For use cases where a block device backed LocalBlobAccess is used to
perform many tiny writes at random offsets, CPU usage is largely
dominated by checksumming. This is because LocalBlobAccess will validate
the integrity of the entire file for every read.

To make such use cases supported, add a special implementation of Buffer
for filesystem.FileReader that doesn't perform any checksum validation.
This obviously comes at the cost of not being able to detect data
corruption. It is up to the user to determine whether that is
acceptable.