Add a FileBackedLocationRecordArray
Allows for the state of a LocalBlobAccess to be stored in a file rather
than in memory. This leads the way to a persistent LocalBlobAccess
blobstore, but also removes the need to store the DigestLocationMap in
memory, which can be limiting to the amount of blobs stored in a
BlockDevice based LocalBlobAccess.

Functionally this implements a FileBackedLocationRecordArray which
essentially creates a file that acts like an array, writing fixed-length
serialised LocationRecords at an offset calculated by the index
multiplied by the length of a record.