Store state for BlockDevice-backed LocalBlobAccess
Allows a LocalBlobAccess configured with a BlockDevice backend to store
the current state, allowing it to be reloaded after a (clean) restart.
The state is stored in a file with an initial integer, expressing the
oldest block ID at time of write, followed by records expressing a
block's ID, whether it's old/current/new and the offset it's placed at.

For example:
  5 ← Oldest block Id
  0,O,123
  0,C,456
  0,N,789 ← Offset in device
  ↑ ↑
  | Whether the block is Old/Current/New
  |
  ID with respect to other blocks of the same type, restarts at 0 for
  each of old/current/new