Add grpc.ClientFactory to stub out gRPC client creation
This allows us to add a DeduplicatingClientFactory that is responsible
for deduplicating clients that are specified in the configuration. If a
bb-storage node has both a CAS and an AC backend that point to the same
host, it's wasteful to create multiple gRPC clients.

Don't tie grpc.ClientFactory into pkg/blobstore/configuration yet; that
code will need some more cleanups before it's capable of using this
well.