Let LocalBlobAccess use a single DigestLocationMap
We needed to use PerInstanceDigestLocationMap to be able to distinguish
between multiple instance names. This is a bit inflexible, because it
doesn't allow people to add new instance names on the spot.

Change this code to be a bit more integral to the rest of the blobstore
API. Let it use StorageType to properly create keys of objects. We then
let the DigestLocation map use hashes of keys to address objects. This
is also a bit more future proof, as it allows us to safely add more
hashing algorithms, etc..