Add the Indirect Content Addressable Storage (ICAS) protocol
The ICAS is similar to the CAS, except that it does not store actual
blob contents. Instead, it stores references to locations where blobs
may be found. This makes it possible to integrate external datasets into
the CAS without actually blowing up the CAS size.

More details can be found in this ADR:
https://github.com/buildbarn/bb-adrs/blob/master/0004-icas.md