Make instance names containing slashes work
The string parsing that's done by ByteStreamServer is absolutely not
designed to take slashes in instance names into account. Instead of just
fixing that, let's go the extra mile and implement all of this in a way
that complies to the REv2 spec.

Because multiple instance names could make parsing of ByteStream URLs
ambiguous, the spec requires that the instance name cannot contain
pathname components that use reserved terms such as "blobs" and
"uploads". We should also add such checks. To ensure we never do this
exactly once per request, introduce an InstanceName type. This type is
not much more than a simple box around a string. It's only possible to
construct instances for valid instance names.

Fixes: https://github.com/buildbarn/bb-storage/pull/70