Let ReadCachingBlobAccess use BlobReplicator
ReadCachingBlobAccess currently applies a replication strategy between
the two backends that is identical to what LocalBlobReplicator does. It
takes little effort to also let this use BlobReplicator.

The advantage of using BlobReplicator here is that we can now use all of
the other BlobReplicator implementations as well. For example,
QueuedBlobReplicator can be used to ensure that replication concurrency
remains bounded.

Let's move all BlobReplicator implementations out of
pkg/blobstore/mirrored, as it is no longer specific to
MirroredBlobAccess. Also move ReadCachingBlobAccess in its own package
to prevent cyclic dependencies.