Add a FileInfoList type to make it easier to sort directory listings
Directory.ReadDir() needs to return elements in sorted order. This is
automatically guaranteed by LocalDirectory, due to it using
Readdirnames(). For custom implementations, it may be necesary to call
sort.Sort() manually.

Let's add a simple implementation of sort.Interface that can be used to
sort a slice of FileInfo objects.