Add stackdriver trace
Basic implementation of stackdriver tracing.
Should we extract common fields (like always sample) into a separate structure?
Should we shove this kind of thing under a tracing or debugging field with the jaeger stuff?
I punted on those for now cause it'd be a breaking config change, but more then happy to rework them