gRPC: Add forward_and_reuse_metadata
We already have the forward_metadata option to specify gRPC headers that
need to be forwarded from an incoming context to an outgoing context.
This can, for example, be used for credential forwarding.

This change adds a new option that is similar to forward_metadata,
except that it caches the headers that are forwarded. These are then
attached to all future RPCs. This option can be useful when running
bb_storage on your own system as a private proxy. It allows you to use
Bazel to inject credentials into the private proxy process. Other
clients (e.g., CAS inspection tools) can then access any backing stores
without separate authentication.