Add InstanceName{Patcher,Trie}
Now that we support instance names that are based on pathnames, let's
start to make actual use of that. Instead of doing literal instance name
matching for scheduler and the 'allow AC updates' configuration options,
we can do prefix matching/rewriting.

To assist us with that, we add a trie (prefix tree) implementation for
routing requests based on a longest prefix match. We also add a simple
'patcher' object that can be used to either trim or add prefixes to
instance names.