Add replication support to ReadFallback
Package move is necessary because //pkg/blobstore cannot depend on
//pkg/blobstore/replication

Otherwise this is just cribbed from Mirrored. Note that backwards
compatibility is maintained in the case where replicator is nil, so
existing configs keep doing the same thing (and not everybody wants
replication here, it's not always the right thing to do).

Signed-off-by: Andrew Suffield <asuffield@gmail.com>