Prefer WaitForReady on grpc clients
This CallOption causes client calls to prefer to block when the
underlying transport is broken (any form of tcp connection issue)
rather than fail the call instantly with Unavailable. Deadlines are
still respected, but this will make grpc wait until the call deadline
is reached if a backend is not responding.

For bazel builds, we consistently prefer to take the latency hit of
waiting for a connection to recover rather than instantly failing the
build when any glitch happens.