Add Rename() and Sync() to filesystem.Directory
We're going to need these calls to be able to add persistency to
LocalBlobAccess. The persistent state is going to be written to a file
on disk. We need to provide tools to be able to overwrite the persistent
state file in an atomic way. This can only be done by adding the ability
to fsync() on both files and directories, and to rename one file on top
of the other.