Move the HTTP endpoint creation into pkg/global
Right now each of the bb_* binaries handrolls its own web server code.
Even worse, binaries like bb_scheduler and bb_browser offer a single
HTTP server for both the metrics, pprof and their web UIs. Let's
decouple this.

This has the advantage that if we ever want to make changes to the
behaviour of the HTTP server of all of our binaries, we can now do this
in one place. It also consolidates the dependency on Gorilla mux.