Make NewInstanceNameFromComponents public
This may be useful in case instance names need to be parsed from path
names on disk. Those already use slash separators. If only the list of
pathname components is provided, it's nice to be able to construct an
instance name without recombining the elements into a pathname string
first.