Enable CI on linux-386
Some people are interested in running Buildbarn on 32-bit systems, such
as the Raspberry Pi. I'm not convinced we should provide automated CI
builds for those platforms (yet). That said, let's enable CI for
linux-386. The advantage of that platform is that we can easily run unit
tests for those, even on a linux-amd64 machine.

This should allow us to more easily find bugs that tend to affect 32-bit
platforms in general. In fact, this change alone already found a bug in
LocalBlobAccess that caused it to perform atomic operations against an
unaligned 64-bit integer.

Related: https://github.com/buildbarn/bb-remote-execution/pull/73