Ensure that CASBlobAccess Get() and Put() calls terminate synchronously
The failure code paths of these functions don't ensure right now that
the gRPC channel is fully terminated. In practice, this doesn't cause
any issues, as the channels get cleaned up later on if the surrounding
context gets cancelled, but this this is done asynchronously.

Ensure that all of this is done synchronously, so that the number of
goroutines remains bounded, and to ensure that the go-grpc-prometheus
metrics better reflect reality.