Move NewGenerator() out of Digest
So far, the Buildbarn codebase is not used as part of a build client.
Only as a build service. This means that the digest API hasn't been
designed to allow you to compute digests of files out of thin air. They
always have to be derived from a parent digest (i.e., the action
digest).

To make it easier for people to use the Buildbarn code as part of a
build client, let's add a new digest.Function type that represents the
tuple (instance name, hashing algorithm).