Add a path resolution library
The Go standard library provides the "path" and "path/filepath"
packages. The former is often a bit too dumb for Buildbarn's needs,
because it processes paths without taking the contents of the underlying
file system into account. For example, it assumes that it's always safe
to remove ".." components. The latter is smarter, but assumes that
resolution needs to happen against the host file system. In the case of
Buildbarn, we often want to traverse in-memory file system
representations.

This change adds a new package, pkg/filesystem/path, that adds abstract
tools for dealing with UNIX pathname strings, which we can use
throughout the rest of the codebase:

- Component: a simple box around a string that enforces that the value
  corresponds to a valid UNIX filename. I've observed that about 20% of
  the code in the Buildbarn tree deals with filenames. In certain places
  we already have filename checks in place, but my suspicion is that
  those aren't sufficient. By using this type, we can be certain that
  invalid filenames don't make it into the lower layers of Buildbarn.

- Resolve(): A function for parsing UNIX pathname strings. It will
  initially calls into a ScopeWalker to notify whether a path is
  absolute or relative. For every comopnent that follows, a call against
  a ComponentWalker is made. This function can also parse ".."
  components and perform symlink expansion if needed.

- RelativeScopeWalker: A simple implementation of ScopeWalker that
  requires that paths are relative. There is no need to add an
  AbsoluteWalker just yet, as REv2 only gives us relative paths. We can
  add it as needed.

This change also adjusts the API of filesystem.Directory to use
path.Component. The nice thing is that we no longer need to do any
filename validation in LocalDirectory, as it's guaranteed that the
components have the right shape.