Perform CI builds for 32- and 64-bit ARM
In order to make the build pass, we have to make some slight changes to
the recently added atomic package. We need to give 32-bit ARM the same
treatment as 386, as 64-bit integers are also not 8-byte aligned. If we
ever need to figure this out for other platforms, we can grep Go's
src/runtime/internal/sys/arch_*.go for Int64Align.

Related to:
- https://github.com/buildbarn/bb-deployments/issues/53
- https://github.com/golang/go/issues/9959