Remove Digest.GetHashXAttrName()
This function was added to be able to place an extended attribute on
CAS-backed files in the FUSE file system containing the hash of a file.
The new client-side FUSE daemon I'm developing to speed up Bazel builds
no longer depends on this feature. It will obtain digests of files
through gRPC.