Extend the path resolution library
In the nearby future, I want to release a client-side FUSE daemon that
people can use to speed up builds that use remote execution. More
details can be found here:

https://github.com/bazelbuild/bazel/pull/12823

This daemon is going to make extensive use of the path resolution
library that we added some time. This commit contains all of the
components that the FUSE daemon is going to need in addition to what's
already there:

- AbsoluteScopeWalker: To enforce that paths are absolute.
- Builder: To construct normalized pathname strings.
- LoopDetectingScopeWalker: To prevent infinite traversal of symlinks.
- VirtualRootScopeWalkerFactory: To simulate resolution of absolute
  paths that start with a prefix that needs to be trimmed.
- Void{Component,Scope}Walker: To perform path resolution without making
  any assumptions about the underlying file system.