Add an API for creating digest.Function objects out of thin air
With the previous incarnation of the client-side FUSE daemon, only files
that are actually backed by the CAS are served through FUSE. This turned
out to be insufficient; we need to place all of bazel-out/ on FUSE to
keep the build happy.

If we're going to do that, we also need to provide digests for files
that are not backed by the CAS; files that were written into the FUSE
file system locally. The problem is that such user-provided files don't
have a hashing algorithm associated with them. The client needs to
provide an explicit hashing algorithm name, so that we can compute the
digest using the right algorithm.

This change adds a proper API for being able to do this. Bazel can now
send the hashing algorithm to the FUSE daemon in the form of an REv2
enum value.