Ensure gRPC client metrics are correct for Execute()
Just like how we had to jump through hoops to get gRPC client metrics to
work properly for ByteStream Read() and Write() calls, we have the same with
Execute(). We see that the gRPC client metrics show a very small number of
in-flight Execute() calls, even though the cluster is completely idle.

Let's solve this by ensuring that Recv() keeps on getting called until
an error is received. This allows the gRPC Prometheus middleware to
properly treat the RPC as being completed.