Make the diagnostics HTTP server configuration more flexible
This option currently enables a web server that provides health checking
endpoints, Prometheus metrics and pprof. In certain cases it is
desirable to only provide health checking. Or health checking and
Prometheus, but not pprof.

Let's turn the settings for the diagnostics HTTP server into a message
that can contain its own set of options.