TLS: Add the ability to override the server name
This option is useful in case you want to connect to an endpoint using
literal IPv4/IPv6 addresses, as opposed to using a DNS entry. The server
may provide a certificate that has a Common Name/Subject Alternative
Name set, but those don't match the network address. This option allows
you to specify which hostname needs to be used for certificate
validation.