InstanceName: Add a function for obtaining the components
We already have NewInstanceNameFromComponents(), which is used by
bb_clientd to convert a sequence of pathname components to an instance
name. Now that we want to let bb_browser create links to paths inside of
bb_clientd, we need the inverse: a function to decompose instance name
strings into an array of components.