Make testutil.RequireEqualProto() a bit weaker
Right now comparison fails in case messages contain anypb.Any fields
that contain maps. This is because the order in which map elements are
stored is not deterministic, and comparisons of anypb.Any only look at
the marshaled contents of the embedded message.

Let's fall back to doing a string comparison on the JSON output in case
proto.Equal() fails.