LocalBlobAccess: Let the AC and ICAS always write to one block
For the CAS we want to let LocalBlobAccess write new objects to multiple
blocks, as this ensures that objects are nicely spread out across all of
storage. This amortizes the cost of refreshing data.

For the AC and ICAS it would also be nice to have this, but doing so
would make it a lot harder to supporting updating of existing entries.
We would need to make sure that all overwrites are performed against the
latest block. This would require us to query the key-location map first
to see if we're performing an overwrite.

Instead of doing that, let's simply restrict LocalBlobAccess, so that it
only supports one "new" block. During initialization of LocalBlobAccess,
there is this special mode where we temporarily increase the number of
"new" blocks to get a better spread of data. This also needs to be
disabled.

Add a BlockListGrowthPolicy helper type that determines the maximum
number of "new" blocks and whether the special initialization mode is
permitted.

Fixes: #118