Add the Initial Size Class Cache
The Initial Size Class Cache, or ISCC for short, is a new Protobuf based
data store. We're going to use this to store outcomes of actions on a
per-command basis. This makes it possible to get insight in execution
times of actions of a given shape.

Because it supports multiple so-called size classes, bb_scheduler will
be able to use this data to make smart decisions about which worker size
is the most optimal for a given action.