Add support for an instance name aware CAS
Right now, all Protobuf based data stores respect instance names, while
the CAS does not. It completely ignores it. This reduces the amount of
storage space needed, but has the downside that inside multi-tenant
environments, there is no way to shield off data.

This change adds a configuration option to LocalBlobAccess to setup an
instance name aware CAS. Objects are still stored once, but have
multiple key-location map entries associated with them.

Instance names are managed hierarchically. This means that any blobs
written using instance name "a" are readable using instance names "a",
"a/b", "a/b/c", etc., but not "", "b", "b/b", etc. In other words, there
is a way to inherit CAS contents from a parent. This is useful for
speeding up builds of derived projects.