Add HierarchicalInstanceNamesBlobAccess
We recently added the hierarchical_instance_names option to
LocalBlobAccess, which can be used to make the CAS keep track of blobs
in a per instance name fashion, taking the hierarchy of instance names
into account. In other words, non-empty instance names will inherit the
contents of their parents (i.e., string prefix matching).

This change adds a special implementation of BlobAccess that can be used
to achieve the same thing for the Action Cache (AC), Initial Size Class
Cache (ISCC) and Indirect Content Addressable Storage (ICAS). We need to
treat these data stores differently, because objects in those data
stores are mutable and may reference objects stored in the CAS. Unlike
the CAS, we cannot implement the policy at the individual storage node
level.

When used for the AC, it is recommended that this decorator is placed
above CompletenessCheckingBlobAccess. That way ActionResult lookups will
continue, even if one or more incomplete results are encountered.