Use grpc.ServiceRegistrar where possible
Modern versions of protoc-gen-go shipped with gRPC will emit
grpc.ServiceRegistrar as part of all generated code, as opposed to using
*grpc.Server:

https://github.com/grpc/grpc-go/commit/9519efffeb5d1897ae8671568871a6d476986524#diff-c765195c37749c9c3591e05f8f0c77858ea5aa093e9b6f3f7a339309f9e507b2

The advantage of this is that gRPC services can be registered against
custom server types. Those include the xDS server type that @Jonpez2 is
trying to use in #120.

This change effectively backports the commit linked above to the older
version of protoc-gen-go that is part of the old Protobuf v1 repository.
Ideally we wouldn't use that compiler at all, but it looks like there
are still several blockers on the rules_go side to switch to the updated
compiler.