pkg/global: Use OpenTelemetry instead of OpenCensus
OpenCensus and OpenTracing have effectively merged to form
OpenTelemetry. With OpenTelemetry approach v1.0.0, now is a good time to
switch to it.

This change only brings in support for Jaeger. Members in the community
are invited to add support for other backends once again.

While there, add a maximum rate sampler, so that bb_* binaries can be
configured in such a way that they only generate a maximum number of
traces.

Fixes: #88 #107