Add a utility function for saving W3C Trace Contexts
This method can be used by InMemoryBuildQueue and BuildClient to
propagate traces from the scheduler to the worker. The regular gRPC
middleware can't be used for this purpose, as it only propagates along a
forward direction. Because workers send RPCs to the scheduler, not the
other way around, this won't work.