Allow Authenticator to wrap Context
When implementing custom Authenticators, it may be useful to attach
metadata to the Context, so that it may be passed down into gRPC
services. This data could, for example, be used for Authorization
purpopses.

This change adjusts the prototype of Authenticate() to return a Context.
While there, move the interceptors into a separate source file and give
them testing coverage.