Allow the diagnostics HTTP server to display OpenTelemetry spans
For larger setups, it becomes infeasible to capture all spans. That
said, sometimes you want to do hands-on debugging to processes and see
what all of the spans in the system are. This change adds an
"/active_spans" endpoint to the diagnostics web server that is capable
of simply displaying all of them.

It may, for example, be used to see whether bb_worker is idle or running
build actions.