Add a function for converting gRPC codes to HTTP status codes
Most of the binaries that contain web UIs often have to convert error
objects to HTTP error pages. Provide a utility function that can be used
to let these pages return the right HTTP status code.

The status codes that are returned by this function are the ones that
are documented as part of the gRPC Code enum.