first cut of spiffe support
Spiffe, which stands for Secure Production Identity Framework for Everyone, is a
"A universal identity control plane for distributed systems "

Spiffe implementors use a specific x509 template as described in
    https://github.com/spiffe/spiffe/blob/main/standards/X509-SVID.md

Services and clients that use x509-SVIDs are able to symetically use a single spiffe-id
to verify the identity of their peers. This is achieved by way of spiffe IDs as described in
    https://github.com/spiffe/spiffe/blob/main/standards/SPIFFE-ID.md

Spiffe is being adopted by major cloud providers and vendors and other open-source projects
such as envoy and consul.