Add ConcurrencyLimitingBlobReplicator
I've observed that Bazel may issue an unrealistic number of concurrent
read requests against storage; 1000 is not uncommon. This causes
individual transfers to go very slow, causing them to time out.

This change adds a new BlobReplicator that compensates for that by
limiting the maximum amount of concurrency. It is intended to be used by
bb_clientd. By setting this up, Bazel is permitted to issue a large
number of concurrent read request against bb_clientd, but bb_clientd
will make sure to limit calls against the remote execution service in
case of cache misses.