Add support for JSON Web Token authentication for gRPC
This change adds a new implementation of Authenticator that is capable
of extracting "Authorization: Bearer ${jwt}" metadata from incoming
requests. It can then parse the JSON Web Token and validate its
signature. If valid, requests may continue.

I initially tried to build this on top of github.com/golang-jwt/jwt, but
ran into several issues. One of the most prominent being that it's hard
to inject timekeeping in this library. It also doesn't provide any
information on how long a certain authentication decision is valid,
meaning it's hard to do reliable caching of tokens. I've worked around
this by adding a pkg/jwt that does all of the heavy lifting for us.

One of the interesting things is that this AuthorizationHeaderParser is
capable of caching tokens. It makes use of pkg/eviction to manage its
cache, meaning that existing Prometheus metrics for eviction sets can
also give stats on token cache hit rates.

Fixes: #48